requirements.txt
Imath==0.0.2
numpy==1.26.0
openai==0.28.0
OpenEXR==1.3.9
Pillow==10.0.1
python-dotenv==1.0.0
tiktoken==0.5.1
websockets==11.0.3
